requirements.txt
# Existing dependencies
openai
anthropic
google-generativeai
huggingface_hub
requests
matplotlib
python-dotenv
fastapi
uvicorn
pydantic

# Multi-user cloud dependencies
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
websockets>=12.0
email-validator>=2.0.0
